wip: move verifier to legacy decorator
