+++ src/application/create-application.test.ts
@@ -1,7 +1,7 @@
 import { create } from '@/application/create-application';
 import { koalaDefaultConfig } from '@/config/default-config';
 import { Get, Route, RouteGroup } from '@/routing';
-import { exclusiveRoutingModeError } from '@/routing/verify-routing-mode';
+import { exclusiveRoutingModeError } from '@/routing/deprecated-decorator/verify-routing-mode';
 import { expect, test } from 'vitest';
 
 test('create app with default config', () => {
