+++ src/routing/deprecated-decorator/verify-routing-mode.test.ts
