+++ src/routing/deprecated-decorator/verify-routing-mode.ts
